Use real time for end condition
Until now, the planned time for the next packet was used to check if the
end time had been reached. This could cause two kinds of problems:

a) If the client could not send packets as fast as requested, the packet
times were still be incremented as if it could, leading to a longer
operating time than requested.

b) Even worse, if the requested packet interval was 0, the next packet
time did not increase, leading to an endless loop.

Getting the real time from the kernel after each packet and using it to
check if the planned end time has been reache solves both problems.
Requesting an inter send time of 0 will now cause the client to send
packets as fast as possible.