Introduce named command line options for timediff.m
Rather than relying on order, options can now be specified in any order,
as long as they are followed by a "--" and than the files. For simply
drawing IAT plots, just specifying the file names is sufficient.

Thus, the new syntax is:
timediff.m [OPTIONS --] FILENAME(S)

Files after the first are ignored for now.