timediff.m: Fix bar at the lower limit of IAT plots
The stairs plot created in eval_iat was missing the left part of the
first bar (left-most of the histogram). Adding an additional data point
one bar width left of the existing one solves the problem. At the same
time the X values are shifted half a bar width to the left to match the
layout created when directly plotting from the hist() function.

The X axis limits now depend solely on the histogram range.