Client: Introduce packet blocks
A packet block contains a list of packet data structures.

The idea is that a generation thread can create blocks, while the
sending thread consumes them. Block creation could be static or dynamic,
and the next pointers allow both linear and circular lists. The normal
case should be a circular list. For infinite operation, the generation
thread must be able to create new blocks at least as fast as the sending
thread consumes them.