Generic interface for packet parameter generators
generator_t stores all information the client thread needs to start a
generator thread. The goal is to have the client set up a generator_t
object for the user's generator of choice, and then run it. The generic
interface ensures proper locking and inter-thread signalling, so
developers don't need to bother with management.

The implementation of this concept is far from complete, but the basic
static generator (all packets have same size and interval) works, adding
other static generators would be simple.