Remote control: Separate configuration option for transmission target
By default, the value of the server option will be used in the fast-tg
command on the client as the server address. When a separate control
network is used, the server might have a different address on the
measurement network. This address can be set with the target option, if
required.