Basic self test
The test script creates a minimal remote control configuration for a
transmission with both sides on localhost and runs it. To get a
meaningful result, errors in the control script are now passed through,
up to the main thread. Additionally, the default_exec path in
luna-control gets adjusted to the proper installation path during build.