luna-control: Support common and relative start times
If the general configuration variable "startup_delay" is set, a common
start time for all connections is calculated as current time right
before the connection handlers are started plus the startup delay. To
work properly, the delay must be long enough to accommodate the SSH
connection setup and LUNA startup times.

Additionally, a connection can have its own delay relative to the common
start time. If desired, this can be configured using the "delay" option
in the connection section.